More fixes.

+++ .drupal-dev/tests/test-workflow.sh
@@ -13,7 +13,7 @@ trap 'last_command=$current_command; current_command=$BASH_COMMAND' DEBUG
 trap '[ "$?" != "0" ] && echo "\"${last_command}\" command failed with exit code $?."' EXIT
 
 CURDIR="$(cd "$(dirname "$(dirname "${BASH_SOURCE[0]}")")/.." && pwd)"
-BUILD_DIR=${BUILD_DIR:-/tmp/drupal-dev-workflow}
+BUILD_DIR=${BUILD_DIR:-/tmp/star_wars}
 DRUPAL_VERSION=${DRUPAL_VERSION:-7}
 VOLUMES_MOUNTED=${VOLUMES_MOUNTED:-1}
 
@@ -27,8 +27,8 @@ step(){
 sync_to_host(){
   export $(grep -v '^#' .env | xargs)
   [ "$VOLUMES_MOUNTED" == "1" ] && return
-  echo "Syncing from $(docker-compose ps -q cli) to $BUILD_DIR"
-  docker cp -L $(docker-compose ps -q cli):/app/. $BUILD_DIR
+  echo "Syncing from $(docker-compose ps -q cli) to ${BUILD_DIR}"
+  docker cp -L $(docker-compose ps -q cli):/app/. ${BUILD_DIR}
 }
 
 # Sync files to container in case if volumes are not mounted from host.
@@ -42,22 +42,25 @@ sync_to_container(){
 echo "==> Starting WORKFLOW tests for Drupal ${DRUPAL_VERSION} in build directory ${BUILD_DIR}"
 
 # Prepare build directory.
-rm -Rf $BUILD_DIR > /dev/null
-mkdir -p $BUILD_DIR
-git archive --format=tar HEAD | (cd $BUILD_DIR && tar -xf -)
-# Special treatment for cases where volumes are not mounted from host.
-[ "$VOLUMES_MOUNTED" != "1" ] && sed -i -e "/###/d" $BUILD_DIR/docker-compose.yml
+rm -Rf ${BUILD_DIR} > /dev/null
+mkdir -p ${BUILD_DIR}
+git archive --format=tar HEAD | (cd ${BUILD_DIR} && tar -xf -)
+# Special treatment for cases where volumes are not mounted from the host.
+if [ "${VOLUMES_MOUNTED}" != "1" ] ; then
+  sed -i -e "/###/d" ${BUILD_DIR}/docker-compose.yml
+  sed -i -e "s/##//" ${BUILD_DIR}/docker-compose.yml
+fi
 
-pushd $BUILD_DIR > /dev/null
+pushd ${BUILD_DIR} > /dev/null
 
 step "Initialise the project"
 printf 'Star Wars\n\n\n\n\nno\n\n\n' | ahoy init
 
 step "Create .env.local file"
-echo FTP_HOST=$DB_FTP_HOST >> .env.local
-echo FTP_USER=$DB_FTP_USER >> .env.local
-echo FTP_PASS=$DB_FTP_PASS >> .env.local
-echo FTP_FILE=db_d${DRUPAL_VERSION}.dist.sql >> .env.local
+echo FTP_HOST=${DB_FTP_HOST} >> .env.local
+echo FTP_USER=${DB_FTP_USER} >> .env.local
+echo FTP_PASS=${DB_FTP_PASS} >> .env.local
+echo FTP_FILE=db_d${DRUPAL_VERSION}.star_wars.sql >> .env.local
 
 step "Add all files to new git repo"
 git init
@@ -72,11 +75,12 @@ ahoy download-db
 step "Build project"
 ahoy build
 sync_to_host
-BUILD_DIR=$BUILD_DIR goss --gossfile $CURDIR/.drupal-dev/tests/goss/goss.build.yml validate
+BUILD_DIR=${BUILD_DIR} goss --gossfile ${CURDIR}/.drupal-dev/tests/goss/goss.build.yml validate
 # @todo: Try moving this before test.
 sync_to_container behat.yml
 sync_to_container phpcs.xml
 sync_to_container tests
+# @todo: Add test that the correct DB was loaded (e.g. CURL and grep for page title).
 
 step "Enable development settings"
 cp docroot/sites/default/default.settings.local.php docroot/sites/default/settings.local.php
@@ -97,23 +101,24 @@ ahoy export-db mydb.sql
 
 step "Run single Behat test"
 ahoy test-behat tests/behat/features/homepage.feature
-[ -z "$(ls -A screenshots)" ] && "Behat screenshots were not created" && exit 1
 sync_to_host
+[ -z "$(ls -A screenshots)" ] && "Behat screenshots were not created" && exit 1
 
 step "Build FE assets"
 echo "\$color-silver-chalice: #ff0000;" >> docroot/themes/custom/star_wars/scss/_variables.scss
 ahoy fed
 sync_to_host
+grep -qv '#ff0000' docroot/themes/custom/star_wars/build/css/star_wars.min.css && "FE assets were not created" && exit 1
 
 step "Re-import DB"
 rm -Rf .data/*
 echo "DB_EXPORT_BEFORE_IMPORT=1" >> .env.local
 ahoy download-db
-ahoy install
+ahoy install-site
 ls .data/db_export_* > /dev/null
 
 step "Clean"
 ahoy clean
-BUILD_DIR=$BUILD_DIR goss --gossfile $CURDIR/.drupal-dev/tests/goss/goss.clean.yml validate
+BUILD_DIR=${BUILD_DIR} goss --gossfile ${CURDIR}/.drupal-dev/tests/goss/goss.clean.yml validate
 
 popd > /dev/null
